+++ dbms/mariadb/src/test/java/is/codion/dbms/mariadb/MariaDBDatabaseTest.java
@@ -31,7 +31,7 @@ import java.sql.SQLException;
 import java.sql.SQLTimeoutException;
 import java.util.ResourceBundle;
 
-import static is.codion.common.db.database.Database.Operation.*;
+import static is.codion.common.db.exception.Operation.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 public class MariaDBDatabaseTest {